Extract code from ConfigPrinter
